requirements.txt
fastapi
uvicorn
langchain
langchain-community
openai
pydantic
python-dotenv
pinecone-client
mysql-connector-python
pandas
